refactor(orchestrator): move app and os loops to seperate files
